Refactored user request code into separate package. WIP for district summary update.
